requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
requests==2.32.3
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3